MALI-026 B8-3 Step 5 (3/5, stages 2-6): exact Money aggregate conversion (§15-16)
Full return-type surgery: financial values are exact Money / per-currency
structures everywhere in the domain/provider layer; double survives only at a
terminal presentation boundary (chart geometry, ratio/percentage, formatted
text). D (defect) sites → 0.

VALUE OBJECTS → Money: CurrencyTotal.expense/income, CardSummary +
CardAccountBreakdownRow totalIn/totalOut, CategorySpend.total, MerchantSpend.total,
DailySpend.total; RecurringCandidate gains a `currency` (its averageAmount stays a
labeled heuristic double, §14).

AGGREGATES → exact minor: every money SUM now sums `amount_minor` (no
CAST AS REAL — an int64 overflow raises, fail-exact). expenseTotalBetween /
incomeTotalBetween / categoryExpenseTotalBetween require a `currency` and return
Money; currencyTotalsBetween / card summaries / breakdowns return per-currency
Money; latestBalanceAfter → Money? (balance_after_minor + row currency);
spentForPlan → Money (plan.currency). Predicates stay centralized in FinancialSql.

D-SITES → Money arithmetic: budget/plan/goal remaining + isOver + spent-vs-limit
comparisons; dashboard headline totals + `saved` + pendingReviewTotal +
monthlyBudgetLimit + totalSpend, all scoped to the active display currency
(selected account currency, else base). Ratios/scores/percentages compute from
exact Money then .toDouble() ONLY at the boundary.

CROSS-CURRENCY: never Money(EGP)+Money(SAR). Every list fold that could mix
currencies (dashboard pending/budgets/goals, budgets_screen roll-ups, goals_screen
totals) filters to the display currency first; per-currency surfaces use
List<CurrencyTotal>/grouped results. homeGoalsProvider stays a full active-goal
LIST (each card shows its own currency — no fold, so no currency filter).

Consistency: dashboard and reports now share the same canonical exact predicate;
category/merchant shares compute numerator/denominator in the same currency.

Guard: tools/check_arch_guard.sh gains check 7 — no money SUM is cast to REAL
(exact integer minor totals). The retired zero-consumer Supabase summary adapter
is an EXPLICIT legacy remote-number seam (class C) in the fromLegacyReal guard.

Tests: exact_aggregate_foundation_test (10, incl. int64 SUM overflow fails-exact);
report_multicurrency_test (SAR/EGP/KWD isolated, currencyTotals never merged,
donut = primary not cross-currency, 0/2/3-decimal); plus the full aggregate/
reporting/dashboard/budget/goal/plan/card/finance/transaction suites migrated to
Money. flutter analyze clean; arch guard PASS; 727 focused tests green. Bulk
mechanical conversion delegated to Codex, reviewed for financial correctness +
gated + committed here. Nothing pushed or deployed.

+++ app/lib/features/accounts/accounts_screen.dart
@@ -217,7 +217,7 @@ class _UnassignedCardTile extends StatelessWidget {
                 ],
               ),
             ),
-            Text(Formatters.amount(net),
+            Text(Formatters.amount(net.toDouble()),
                 style: AppTypography.caption(c.textLight)),
             const SizedBox(width: 6),
             Icon(Icons.chevron_left, color: c.textLight, size: 20),
